Keep tk-ssh-agent running
Keep the agent running all the time and deal with SIGHUP from udev rules
that tells us something has happened with the device. Reconnect and load
device app as necessary.

Pinentry is used to get USS if start with --uss. The state of --uss and
--uss-file arguments are kept in the signer. Ask for USS again with
pinentry or read the USS file again when the stick has been re-inserted.

Co-authored-by: Daniel Lublin <daniel@lublin.se>

Signed-off-by: Michael Cardell Widerkrantz <mc@tillitis.se>
Signed-off-by: Daniel Lublin <daniel@lublin.se>