Change NameVersion to an ASCII array - BREAKS CDI!
We now receive an ASCII array from firmware for NAME0 and NAME1
instead of a little endian uint32. See firmware

commit a0974a2c0b0ec4f2444ebb8710638ed4d9386c83

Since we use the same tk1.go:GetNameVersion() when talking to
signerapp from tk-ssh-agent, change what signerapp sends to an ASCII
array too. NOTE that this changes the measurement of the signerapp and
thus changes CDI!

Signed-off-by: Michael Cardell Widerkrantz <mc@tillitis.se>