Licenses: Make it clearer what license we release under
Update the README with license text.

For detection purposes we use isolate all programs under different
licenses in their own directory. There will be a LICENSE files in
every such directory and a single LICENSE file in the top directory
covering most of the project.

The LICENSES directory disappears and the spdx-ensure script is moved
to tools.