Add script for running TKey/QEMU container with latest firmware
And the dockerfile for building the image. We're using the somewhat
treacherous "latest" tag to allow script to easily update the image for
the user. Need to remember to push new images (with a plain number tag)
also to "latest" tag.

Include firmware built with both -DNOCONSOLE and built without it. Will
be useful later.

Signed-off-by: Daniel Lublin <daniel@lublin.se>