Use external device libs
We move all device libraries to their own repo:

  https://github.com/tillitis/tkey-libs

A new LIBDIR is introduced in Makefile that should point to the path
where tkey-libs is cloned. It's settable on command line but with a
sensible default of a path next to where this repo is cloned.

Github Actions is changed to clone and build tkey-libs before building
here.