Load different device apps depending on TKey Model
The correct app is loaded depending on what app type we identify with
internal function signer.go:identifyAppType(). If we can't identify
what app to use, notify user.

The signer device app used when Acrab or Bellatrix models are detected
is unchanged. This means that already used identities are unchanged.

For development purposes the Castor device app is used when a device
with UDI PID 0 is detected.

Co-authored-by: Michael Cardell Widerkrantz <mc@tillitis.se>